EL-3170 - Tree Grid Hybrid
