--- docs/app/pages/components/components-sections/tree-view/tree-grid-ng1/wrapper/tree-grid-wrapper.directive.html
@@ -1,34 +1,0 @@
-<treegrid data="vm.data" columns="vm.columns" selected="vm.selected" current-row="vm.currentRow" options="vm.options"></treegrid>
-
-<p class="m-t-md">Selected items: <span ng-repeat="item in vm.selected" ng-bind="item.title + ($last ? '.' : ', ')"></span><em ng-if="vm.selected.length === 0">none</em></p>
-
-<div class="row uxd-customize-example">
-    <div class="col-md-12">
-        <accordion>
-            <accordion-group heading="Customize Example...">
-                <div class="row uxd-customize-row">
-                    <div class="col-md-4 col-sm-12">
-                        <checkbox ng-model="vm.enableDisplayPanel">Display Panel</checkbox>
-                    </div>
-                </div>
-                <div class="row uxd-customize-row">
-                    <div class="col-md-4 col-sm-12">
-                        <checkbox ng-model="vm.options.select.row">select.row</checkbox>
-                    </div>
-                    <div class="col-md-4 col-sm-12">
-                        <checkbox ng-model="vm.options.select.check">select.check</checkbox>
-                    </div>
-                    <div class="col-md-4 col-sm-12">
-                        <checkbox ng-model="vm.options.select.selectChildren" ng-disabled="vm.options.select.row">select.selectChildren</checkbox>
-                    </div>
-                </div>
-                <div class="row uxd-customize-row">
-                    <div class="col-md-6 col-sm-12">
-                        <label for="select-rowClass">select.rowClass</label>
-                        <input type="text" name="select-rowClass" class="form-control" ng-model="vm.options.select.rowClass">
-                    </div>
-                </div>
-            </accordion-group>
-        </accordion>
-    </div>
-</div>
