+++ docs/app/pages/components/components-sections/tree-view/tree-view.module.ts
@@ -1,22 +1,24 @@
 import { CommonModule } from '@angular/common';
-import { NgModule, ComponentFactoryResolver, NO_ERRORS_SCHEMA } from '@angular/core';
+import { ComponentFactoryResolver, Injector, NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
+import { TreeModule } from 'angular-tree-component';
+import { AccordionModule } from 'ngx-bootstrap/accordion';
+import { TabsModule } from 'ngx-bootstrap/tabs';
+import { CheckboxModule, FocusIfModule, HybridModule } from '../../../../../../src/index';
 import { DocumentationComponentsModule } from '../../../../components/components.module';
-import { ResolverService, DocumentationPage } from '../../../../services/resolver/resolver.service';
 import { DocumentationCategoryComponent } from '../../../../components/documentation-category/documentation-category.component';
-
-import { ComponentsTreeViewNg1Component } from './tree-view-ng1/tree-view-ng1.component';
-import { ComponentsTreeViewCompanionViewNg1Component } from './tree-view-companion-view-ng1/tree-view-companion-view-ng1.component';
+import { DocumentationPage, ResolverService } from '../../../../services/resolver/resolver.service';
+import { WrappersModule } from '../../../../wrappers/wrappers.module';
 import { ComponentsTreeGridAsynchronousLoadingNg1Component } from './tree-grid-asynchronous-loading-ng1/tree-grid-asynchronous-loading-ng1.component';
 import { ComponentsTreeGridNg1Component } from './tree-grid-ng1/tree-grid-ng1.component';
-import { TabsModule } from 'ngx-bootstrap/tabs';
-import { WrappersModule } from '../../../../wrappers/wrappers.module';
+import { ComponentsTreeViewCompanionViewNg1Component } from './tree-view-companion-view-ng1/tree-view-companion-view-ng1.component';
+import { ComponentsTreeViewCustomNodeComponent } from './tree-view-custom-node/tree-view-custom-node.component';
+import { TreeViewService } from './tree-view-custom-node/tree-view-custom-node.service';
+import { ComponentsTreeViewNg1Component } from './tree-view-ng1/tree-view-ng1.component';
 import { ComponentsTreeViewComponent } from './tree-view/tree-view.component';
 
-import { TreeModule } from 'angular-tree-component';
-import { CheckboxModule, FocusIfModule } from '../../../../../../src/index';
-import { TreeViewService } from './tree-view-custom-node/tree-view-custom-node.service';
-import { ComponentsTreeViewCustomNodeComponent } from './tree-view-custom-node/tree-view-custom-node.component';
+
 
 const SECTIONS = [
     ComponentsTreeViewNg1Component,
@@ -39,12 +41,15 @@ const ROUTES = [
 
 @NgModule({
     imports: [
+        AccordionModule.forRoot(),
         CommonModule,
         WrappersModule,
         TabsModule,
         TreeModule,
+        FormsModule,
         CheckboxModule,
         FocusIfModule,
+        HybridModule,
         DocumentationComponentsModule,
         RouterModule.forChild(ROUTES)
     ],
@@ -52,7 +57,17 @@ const ROUTES = [
     declarations: SECTIONS,
     entryComponents: SECTIONS,
     providers: [
-        TreeViewService
+        TreeViewService,
+        {
+            provide: '$templateCache',
+            useFactory: (injector: Injector) => injector.get('$templateCache'),
+            deps: ['$injector']
+        },
+        {
+            provide: '$displayPanel',
+            useFactory: (injector: Injector) => injector.get('$displayPanel'),
+            deps: ['$injector']
+        }
     ]
 })
 export class ComponentsTreeViewModule {
